requirements-dev.txt
# Dev / CI tooling
pre-commit==4.5.1
ruff==0.14.10
mypy==1.19.1
types-requests==2.32.4.20250913
pytest-cov==7.0.0
pip-audit==2.10.0